+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.35] - 2016-08-03
+### Added
+- tile widget method newTileGrid() supports "getWidgetProperty and child property"
+
+### Change
+- fixed bug in slider widget
+
 ## [0.1.34] - 2016-08-03
 ### Added
 - tile widget method newTileGrid() supports "touch points". It's enabled in the demo. See tile.lua for an example and to view demo, click on last circle button in main scene. Next version will support dynamic changing tiles.